Upgrade to MvvmCross 6.2 Update packaging
