+++ MvvX.Plugins.CryptoTools.Platform.Tests/MvvX.Plugins.CryptoTools.Tests.csproj
@@ -1,0 +1,23 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
+
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.8.0" />
+    <PackageReference Include="MSTest.TestAdapter" Version="1.3.2" />
+    <PackageReference Include="MSTest.TestFramework" Version="1.3.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="ComputeHash\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MvvX.Plugins.CryptoTools.Platform\MvvX.Plugins.CryptoTools.csproj" />
+  </ItemGroup>
+
+</Project>
