+++ MvvX.Plugins.CryptoTools.Platform.Tests/UnitTest1.cs
@@ -1,0 +1,13 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace MvvX.Plugins.CryptoTools.Platform.Tests
+{
+    [TestClass]
+    public class UnitTest1
+    {
+        [TestMethod]
+        public void TestMethod1()
+        {
+        }
+    }
+}
