+++ MvvX.Plugins.CryptoTools.Platform/HashType.cs
