+++ MvvX.Plugins.CryptoTools.Platform/ICryptoTools.cs
