--- MvvX.Plugins.CryptoTools.Platform/MvvX.Plugins.CryptoTools.Platform.csproj
@@ -1,12 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard1.6</TargetFramework>
-    <GeneratePackageOnBuild>False</GeneratePackageOnBuild>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\MvvX.Plugins.CryptoTools\MvvX.Plugins.CryptoTools.csproj" />
-  </ItemGroup>
-
-</Project>
