+++ MvvX.Plugins.CryptoTools.Platform/SymetricAlgorithmType.cs
