--- MvvX.Plugins.CryptoTools.Wpf/Plugin.cs
@@ -1,14 +1,0 @@
-﻿using MvvmCross.Platform;
-using MvvmCross.Platform.Plugins;
-using MvvX.Plugins.CryptoTools.Platform;
-
-namespace MvvX.Plugins.CryptoTools.Wpf
-{
-    public class Plugin : IMvxPlugin
-    {
-        public void Load()
-        {
-            Mvx.RegisterSingleton<ICryptoTools>(new PlatformCryptoTools());
-        }
-    }
-}
