--- MvvX.Plugins.CryptoTools.Wpf/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="MvvmCross.Platform" version="4.2.2" targetFramework="net45" />
-</packages>
