--- MvvX.Plugins.Cryptotools/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="MvvmCross.Platform" version="4.2.2" targetFramework="portable45-net45+win8+wpa81" />
-  <package id="Xamarin.Forms" version="2.3.4.224" targetFramework="portable45-net45+win8+wpa81" />
-</packages>
